requirements.txt
astroid==2.4.2
certifi==2020.11.8
chardet==3.0.4
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
pkg-resources==0.0.0
psycopg2-binary==2.8.6
pylint==2.6.0
requests==2.25.0
six==1.15.0
toml==0.10.2
typed-ast==1.4.1
urllib3==1.26.2
wrapt==1.12.1
